Make `MultiSearchItem.status` optional (#660)
* Add failing msearchTemplate integration test


* Make `MultiSearchItem.status` optional


* Add changelog entry


* spotlessApply


---------

(cherry picked from commit d2916e42f7c1af931192bb48a6079948d6af1535)

+++ java-client/src/test/java/org/opensearch/client/opensearch/json/jackson/JacksonJsonpParserTest.java
@@ -186,7 +186,7 @@ public class JacksonJsonpParserTest extends ModelTestCase {
 
         assertEquals(2, response.responses().size());
         assertEquals(404, response.responses().get(0).failure().status());
-        assertEquals(200, response.responses().get(1).result().status());
+        assertEquals((Integer) 200, response.responses().get(1).result().status());
     }
 
     public static class AttributedJacksonJsonpMapper extends JacksonJsonpMapper {
